wip: raster
