--- examples/demos/raster/index.ts
@@ -1,6 +1,0 @@
-export { MapRender as dem } from './dem';
-export { MapRender as image } from './image';
-export { MapRender as image_box } from './image_coord';
-export { MapRender as multiband } from './multiband';
-export { MapRender as ndvi } from './ndvi';
-export { MapRender as tiff } from './tiff';
