--- examples/demos/template.ts
@@ -1,15 +1,0 @@
-import { Scene } from '@antv/l7';
-import * as allMap from '@antv/l7-maps';
-import type { RenderDemoOptions } from '../types';
-
-export function MapRender(options: RenderDemoOptions) {
-  const scene = new Scene({
-    id: 'map',
-    renderer: options.renderer,
-    map: new allMap[options.map]({
-      style: 'light',
-      center: [121.434765, 31.256735],
-      zoom: 14.83,
-    }),
-  });
-}
