+++ examples/demos_next/raster/index.ts
@@ -1,0 +1,6 @@
+export { dem } from './dem';
+export { image } from './image';
+export { imageBox } from './imageBox';
+export { multiband } from './multiband';
+export { ndvi } from './ndvi';
+export { tiff } from './tiff';
